+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Dry-run result output improvements [1123](https://github.com/paritytech/cargo-contract/pull/1123)
+
 ## [3.0.1]
 
 ### Fixed